Add files via upload
In this project, we will assume the role of a Data Analyst recently hired by a global IT and business consulting firm renowned for its IT solutions and experienced team of IT consultants. To stay competitive and keep up with evolving technologies, our organization frequently analyzes data to identify future skill requirements.

As a Data Analyst, we will support this initiative by collecting data from various sources and identifying trends for this year's report on emerging skills. Our primary task is to gather information on the most in-demand programming skills from job postings, training portals, and surveys.

After collecting sufficient data, we will analyze it to uncover insights and trends. Each step of the data analysis process is detailed in different notebooks within this repository, along with additional files necessary for comprehension.